:recycle: refactor(period-selector): Refactor and code cleanup
